+++ src/api/utils.ts
@@ -417,3 +417,17 @@ export function printTree(treeItem: TreeItem, level: number, color: (x: string) 
         printTree(child, level + 1, color);
     });
 }
+
+export function removeTrailingSlash(url: string): string {
+    // billboard ends with /
+    // console.log(url);
+    if (url === "/api/assetmanagement/v3/") return url;
+
+    if (url.includes("?")) {
+        const parts = url.split("?");
+        parts[0] = parts[0].replace(/\/+$/, "");
+        return parts.join("?");
+    } else {
+        return url.replace(/\/+$/, "");
+    }
+}