+++ test/mindconnect-agent-3.spec.ts
@@ -799,8 +799,8 @@ describe("MindConnectApi Version 3 Agent (SHARED_SECRET)", () => {
         }
 
         const agentSdk = agent.Sdk();
-        const billboard = await retry(5, () => agentSdk.GetAssetManagementClient().GetBillboard());
-        billboard.should.not.be.undefined;
+        const assets = await retry(5, () => agentSdk.GetAssetManagementClient().GetAssets());
+        assets.should.not.be.undefined;
         const assetTypes = await retry(5, () => agentSdk.GetAssetManagementClient().GetAspectTypes());
         assetTypes.should.not.be.undefined;
     });